refactor(Comparable.php): move to core
The interface will not be used for customization so should not be in the Api namespace.

Issue #13, Resolves #13
