+++ tests/__comparables__/NestedXSSFix.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Efortmeyer\Polar\Tests\Comparables;
 
-use Efortmeyer\Polar\Api\Comparable;
+use Efortmeyer\Polar\Core\Comparable;
 
 class NestedXSSFix implements Comparable
 {
