+++ tests/__extensions__/PolarTestCaseExtension.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Efortmeyer\Polar\Tests\Extensions;
 
-use Efortmeyer\Polar\Api\Comparable;
+use Efortmeyer\Polar\Core\Comparable;
 use PHPUnit\Framework\TestCase;
 
 class PolarTestCaseExtension extends TestCase
